Add @http.root() function

+++ src/parameter-binder/baseclasses.ts
@@ -11,7 +11,7 @@ export interface BinderCommand {
 export function autoConvert(source: string) {
     if(typeof source == "undefined" || source == null) return;
     if (!isNaN(+source))
-        return parseInt(source);
+        return parseFloat(source);
     else if (source.toLowerCase() === "true" || source.toLowerCase() === "false")
         return source.toLowerCase() === "true";
     else return source;
